Moved config-manager.cjs from utils to lib
