+++ website/server/lib/ServerInstance.cjs
@@ -15,7 +15,7 @@ const {DiscordBot} = require("./DiscordBot.cjs");
 const DiscordBotList = require("@server-lib/DiscordBotList.cjs");
 const {customLog} = require("@server-utils/custom-utils.cjs");
 const ServerManagerList = require("@server-lib/ServerManagerList.cjs");
-const {ConfigManager, ConfigTypes} = require("@server-utils/config-manager.cjs");
+const {ConfigManager, ConfigTypes} = require("@server-utils/ConfigManager.cjs");
 const SocketEvents = require("@server-lib/SocketEvents.cjs");
 const {getBoardByPID} = require("@server-utils/arduino-utils.cjs");
 const ServerList = require("@server-lib/ServerList.cjs");
